Wrap up changes, use combined def of reporter
