+++ breadbox/breadbox/main.py
@@ -1,14 +1,12 @@
-from .startup import create_app, ensure_directories_exist, GCPExceptionReporter
+from .startup import create_app, ensure_directories_exist
 from .celery_task.utils import create_celery
-from .config import get_settings, Settings
+from .config import get_settings
 from breadbox.api.dependencies import get_user
 from fastapi import Request, status
-from fastapi.exceptions import RequestValidationError
 from fastapi.exception_handlers import (
-    request_validation_exception_handler,
     http_exception_handler,
 )
-from .logging_config import configure_logging
+from .logging import configure_logging, GCPExceptionReporter
 
 configure_logging()
